toaster: blocks for custom/highlighted navigation and breadcrumb links
Implement the navigation links as blocks so that each page can override and have
its respective link appear highlighted. Make the build breadcrumb a block so that
it is customizable to not be a link for the dashboard page. Reorder the page headers
to be consistent order for extends, projecttags, localbreadcrumb, nav-links.

[YOCTO #5916]
[YOCTO #4258]

Signed-off-by: David Reyna <David.Reyna@windriver.com>